+++ lib/easypost/beta/payment_refund.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+# PaymentRefund objects represent a refund of a payment.
+class EasyPost::Beta::PaymentRefund < EasyPost::Resource
+end